+++ react/src/blueTheme.ts
@@ -7,6 +7,7 @@
  **/
 
 import { ThemeOptions } from '@material-ui/core/styles/createMuiTheme';
+import MUIShadows from '@material-ui/core/styles/shadows';
 import { typography, createSimpleLightPalette as createSimplePalette } from './shared';
 import * as PXBColors from '@pxblue/colors';
 import Color from 'color';
@@ -48,6 +49,18 @@ const WhiteText = PXBColors.white[50];
 const BlackBorder = PXBColors.black[500];
 const Spacing = 8;
 
+const ShadowBaseColorArray = Color(PXBColors.black[500]).rgb().array();
+const MUIShadowCopy: string[] = MUIShadows.slice();
+MUIShadowCopy.splice(0, 1);
+const TunedShadow = new Array<string>(24);
+TunedShadow.fill('');
+MUIShadowCopy.forEach((shadow, index): void => {
+    TunedShadow[index] = shadow.replace(
+        /rgba\(0,0,0/g,
+        `rgba(${ShadowBaseColorArray[0]},${ShadowBaseColorArray[1]},${ShadowBaseColorArray[2]}`
+    );
+});
+
 /*
     Refer to https://material-ui.com/customization/default-theme/ for a list of properties that are available
     to customize in our themes. These have changed periodically from version to version of Material UI.
@@ -67,6 +80,8 @@ export const blueTheme: ThemeOptions = {
         text: ThemeColors.text,
         action: ThemeColors.action,
     },
+    // @ts-ignore
+    shadows: ['none'].concat(TunedShadow),
     overrides: {
         // APP BAR OVERRIDES
         MuiAppBar: {
@@ -85,7 +100,6 @@ export const blueTheme: ThemeOptions = {
                 },
             },
         },
-
         // AVATAR OVERRIDES
         MuiAvatar: {
             colorDefault: {